+++ SciQLop/core/mime/types.py
@@ -2,4 +2,5 @@ from SciQLopPlots import ProductsModel
 TIME_RANGE_MIME_TYPE = "application/x.sciqlop.time-range"
 PRODUCT_LIST_MIME_TYPE = ProductsModel.mime_type()
 CATALOG_LIST_MIME_TYPE = "application/x.sciqlop.catalog-list"
+EVENT_LIST_MIME_TYPE = "application/x.sciqlop.event-list"